+++ graphile/graphile-query/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.4](https://github.com/constructive-io/constructive/compare/graphile-query@4.5.3...graphile-query@4.5.4) (2026-03-13)
+
+**Note:** Version bump only for package graphile-query
+
+## [4.5.3](https://github.com/constructive-io/constructive/compare/graphile-query@4.5.2...graphile-query@4.5.3) (2026-03-13)
+
+**Note:** Version bump only for package graphile-query
+
 ## [4.5.2](https://github.com/constructive-io/constructive/compare/graphile-query@4.5.1...graphile-query@4.5.2) (2026-03-12)
 
 **Note:** Version bump only for package graphile-query