+++ graphql/test-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.4](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.5.3...@constructive-io/test-codegen-app@1.5.4) (2026-03-13)
+
+**Note:** Version bump only for package @constructive-io/test-codegen-app
+
+## [1.5.3](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.5.2...@constructive-io/test-codegen-app@1.5.3) (2026-03-13)
+
+**Note:** Version bump only for package @constructive-io/test-codegen-app
+
 ## [1.5.2](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.5.1...@constructive-io/test-codegen-app@1.5.2) (2026-03-12)
 
 **Note:** Version bump only for package @constructive-io/test-codegen-app