+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.6.1](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.6.0...@constructive-io/cli@7.6.1) (2026-03-13)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+# [7.6.0](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.5.2...@constructive-io/cli@7.6.0) (2026-03-13)
+
+### Features
+
+- **codegen:** add config command, helpers.ts generation, and collision handling ([4273a61](https://github.com/constructive-io/constructive/commit/4273a618989f6eb55ae120221b97167074de04ff)), closes [constructive-io/dev-utils#70](https://github.com/constructive-io/dev-utils/issues/70)
+
 ## [7.5.2](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@7.5.1...@constructive-io/cli@7.5.2) (2026-03-12)
 
 **Note:** Version bump only for package @constructive-io/cli